backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.25
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv==1.0.0
openai>=1.12.0
numpy>=1.26.0
pandas>=2.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
